+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### change on [04/26/2024]
+
+- add dynamicRanges, celltype detection, neighborhoods analysis scripts.
+
 ### change on [04/25/2024]
 
 - update the documentations and scripts.